Split out browser tests to own file
