+++ tests/run.sh
@@ -16,9 +16,9 @@ usage() {
 case $1 in
 start)
     wget -q --tries=180 --retry-connrefused --waitretry=1 -O /dev/null $SERVER || (echo "Can't reach $SERVER" && exit 1)
-    http --check-status $SERVER/__heartbeat__
-    http POST $SERVER/__flush__
-    pytest --driver Remote --capability browserName firefox --base-url $SERVER/admin --verify-base-url integration_test.py --server $SERVER
+    http -q --check-status $SERVER/__heartbeat__
+    pytest integration_test.py --server $SERVER
+    pytest --driver Remote --capability browserName firefox --base-url $SERVER/admin --verify-base-url browser_test.py --server $SERVER
     ;;
 *)
     exec "$@"
